--- src/test/resources/jsonExamples/payloads/singleCasePronouncementDate.json
@@ -1,6 +1,0 @@
-{
-  "DecreeNisiGrantedDate": "2000-01-01",
-  "DAEligibleFromDate": "2000-02-13",
-  "PronouncementJudge": "District Judge",
-  "hearingDate": "2020-03-20"
-}
